Storage - Fix download problem

+++ apphub-lib/apphub-lib-common_util/src/test/java/com/github/saphyra/apphub/lib/common_util/ValidationUtilTest.java
@@ -89,7 +89,7 @@ public class ValidationUtilTest {
 
     @Test
     public void atLeast_null() {
-        Throwable ex = catchThrowable(() -> ValidationUtil.atLeast(null, 10, FIELD));
+        Throwable ex = catchThrowable(() -> ValidationUtil.atLeast((Integer) null, 10, FIELD));
 
         ExceptionValidator.validateInvalidParam(ex, FIELD, "must not be null");
     }
