+++ apphub-service/apphub-service-platform/apphub-service-platform-storage/src/main/java/com/github/saphyra/apphub/service/platform/storage/service/store/StoredFileFactory.java
@@ -17,7 +17,7 @@ class StoredFileFactory {
     private final IdGenerator idGenerator;
     private final DateTimeUtil dateTimeUtil;
 
-    public StoredFile create(UUID userId, String extension, Integer size) {
+    public StoredFile create(UUID userId, String extension, Long size) {
         return StoredFile.builder()
             .storedFileId(idGenerator.randomUuid())
             .userId(userId)
